requirements.txt
flask==2.0.1
PyPDF2==3.0.1
pdfplumber==0.7.4
python-docx==0.8.11
Werkzeug==2.0.1 